[fragment] Generalize code
