+++ yt_dlp/downloader/fragment.py
@@ -4,6 +4,12 @@ import os
 import time
 import json
 
+try:
+    import concurrent.futures
+    can_threaded_download = True
+except ImportError:
+    can_threaded_download = False
+
 from .common import FileDownloader
 from .http import HttpFD
 from ..utils import (
@@ -112,11 +118,15 @@ class FragmentFD(FileDownloader):
             return False, None
         if fragment_info_dict.get('filetime'):
             ctx['fragment_filetime'] = fragment_info_dict.get('filetime')
-        down, frag_sanitized = sanitize_open(fragment_filename, 'rb')
+        ctx['fragment_filename_sanitized'] = fragment_filename
+        return True, self._read_fragment(ctx)
+
+    def _read_fragment(self, ctx):
+        down, frag_sanitized = sanitize_open(ctx['fragment_filename_sanitized'], 'rb')
         ctx['fragment_filename_sanitized'] = frag_sanitized
         frag_content = down.read()
         down.close()
-        return True, frag_content
+        return frag_content
 
     def _append_fragment(self, ctx, frag_content):
         try:
@@ -304,3 +314,111 @@ class FragmentFD(FileDownloader):
             'tmpfilename': tmpfilename,
             'fragment_index': 0,
         })
+
+    def download_and_append_fragments(self, ctx, fragments, pack_func=None):
+        fragment_retries = self.params.get('fragment_retries', 0)
+        skip_unavailable_fragments = self.params.get('skip_unavailable_fragments', True)
+        test = self.params.get('test', False)
+        if not pack_func:
+            pack_func = lambda frag_content, _: frag_content
+
+        def download_fragment(fragment):
+            i = fragment['index']
+            frag_index = fragment['frag_index']
+            frag_url = fragment['url']
+
+            ctx['fragment_index'] = frag_index
+            headers = info_dict.get('http_headers', {})
+            byte_range = fragment.get('byte_range')
+            if byte_range:
+                headers['Range'] = 'bytes=%d-%d' % (byte_range['start'], byte_range['end'] - 1)
+
+            # Never skip the first fragment
+            fatal = i == 0 or not skip_unavailable_fragments
+            count = 0
+            while count <= fragment_retries:
+                try:
+                    success, frag_content = self._download_fragment(ctx, frag_url, info_dict, headers)
+                    if not success:
+                        return False, frag_index
+                    break
+                except compat_urllib_error.HTTPError as err:
+                    # Unavailable (possibly temporary) fragments may be served.
+                    # First we try to retry then either skip or abort.
+                    # See https://github.com/ytdl-org/youtube-dl/issues/10165,
+                    # https://github.com/ytdl-org/youtube-dl/issues/10448).
+                    count += 1
+                    if count <= fragment_retries:
+                        self.report_retry_fragment(err, frag_index, count, fragment_retries)
+                except DownloadError:
+                    # Don't retry fragment if error occurred during HTTP downloading
+                    # itself since it has own retry settings
+                    if not fatal:
+                        break
+                    raise
+
+            if count > fragment_retries:
+                if not fatal:
+                    return False, frag_index
+                ctx['dest_stream'].close()
+                self.report_error('Giving up after %s fragment retries' % fragment_retries)
+                return False, frag_index
+
+            decrypt_info = fragment.get('decrypt_info')
+            if decrypt_info and decrypt_info['METHOD'] == 'AES-128':
+                iv = decrypt_info.get('IV') or compat_struct_pack('>8xq', fragment['media_sequence'])
+                decrypt_info['KEY'] = decrypt_info.get('KEY') or self.ydl.urlopen(
+                    self._prepare_url(info_dict, info_dict.get('_decryption_key_url') or decrypt_info['URI'])).read()
+                # Don't decrypt the content in tests since the data is explicitly truncated and it's not to a valid block
+                # size (see https://github.com/ytdl-org/youtube-dl/pull/27660). Tests only care that the correct data downloaded,
+                # not what it decrypts to.
+                if not test:
+                    frag_content = AES.new(
+                        decrypt_info['KEY'], AES.MODE_CBC, iv).decrypt(frag_content)
+
+            return frag_content, frag_index
+
+        def append_fragment(frag_content, frag_index):
+            if not frag_content:
+                fatal = frag_index == 1 or not skip_unavailable_fragments
+                if not fatal:
+                    self.report_skip_fragment(frag_index)
+                    return True
+                else:
+                    ctx['dest_stream'].close()
+                    self.report_error(
+                        'fragment %s not found, unable to continue' % frag_index)
+                    return False
+            self._append_fragment(ctx, pack_func(frag_content, frag_index))
+            return True
+
+        max_workers = self.params.get('concurrent_fragment_downloads', 1)
+        if can_threaded_download and max_workers > 1:
+            self.report_warning('The download speed shown is only of one thread. This is a known issue')
+            _download_fragment = lambda f: (f, download_fragment(f)[1])
+            with concurrent.futures.ThreadPoolExecutor(max_workers) as pool:
+                futures = [pool.submit(_download_fragment, fragment) for fragment in fragments]
+                done, not_done = concurrent.futures.wait(futures, timeout=0)  # timeout must be 0 to return instantly
+                try:
+                    while not_done:
+                        # Check every 1 second for KeyboardInterrupt
+                        freshly_done, not_done = concurrent.futures.wait(not_done, timeout=1)
+                        done |= freshly_done
+                except KeyboardInterrupt:
+                    for future in not_done:
+                        future.cancel()
+                    concurrent.futures.wait(not_done, timeout=None)  # timeout must be none to cancel
+                    raise KeyboardInterrupt
+
+            for fragment, frag_index in map(lambda x: x.result(), futures):
+                result = append_fragment(self._read_fragment(fragment), frag_index)
+                if not result:
+                    return False
+        else:
+            for fragment in fragments:
+                frag_content, frag_index = download_fragment(fragment)
+                result = append_fragment(frag_content, frag_index)
+                if not result:
+                    return False
+
+        self._finish_frag_download(ctx)
